refactor: refactor notFound page and useNavigator hook
